requirements.txt
streamlit>=1.33,<2
streamlit-autorefresh>=1.0,<2
MLB-StatsAPI>=1.6,<2
requests>=2.28
bcrypt>=4.0,<5
